refactor(unit_tests): move nemesis tests to unit_tests/unit/nemesis/
Scaffold the unit/integration test directory split so that PR #14372
(network nemesis extraction) can be backported without conflicts.

Moves unit_tests/nemesis/ -> unit_tests/unit/nemesis/ and updates
all imports from unit_tests.nemesis to unit_tests.unit.nemesis.
